Added `showSlotEndTime` column to `EventType` table in database
After modifying the `schema.prisma` file in `packages/prisma`, ran `yarn
db-migrate` to automatically create the migration and `yarn
generate-schemas` to make sure the formatting of the `.prisma` file is
correct and that the `zod` typedefs are generated in
`packages/prisma/client/`.

To confirm the changes, I then ran `yarn db-studio` and visited
`localhost:5555`, then navigated to the view for the `EventType` table
and made sure the `showSlotEndTime` is there.